Testing the code from internet
Seems like pcf8574 doesnt send ack back to master.
